requirements.txt
customtkinter
requests
pan-os-python 